chore: remove dead code and consolidate duplicate BoardName type
Remove ~430 lines of dead code: unused getClimbName.ts, unfinished
climb-info/ component directory, superseded URL utility functions
(constructClimbList, constructClimbSearchUrl), and ~300 lines of
speculative type definitions in aurora-rest-client/types.ts. Consolidate
duplicate BoardName/AuroraBoardName types to a single source of truth
in aurora/types.ts.
