--- packages/web/app/components/climb-info/climb-info-drawer.module.css
@@ -1,9 +1,0 @@
-.desktopOnly {
-  display: none;
-}
-
-@media (min-width: 768px) {
-  .desktopOnly {
-    display: block;
-  }
-}
