--- packages/web/app/components/climb-info/climb-info-drawer.tsx
@@ -1,15 +1,0 @@
-'use client';
-
-import React from 'react';
-import ClimbInfo from './climb-info';
-import styles from './climb-info-drawer.module.css';
-
-const ClimbInfoColumn = () => {
-  return (
-    <div className={styles.desktopOnly}>
-      <ClimbInfo />
-    </div>
-  );
-};
-
-export default ClimbInfoColumn;
