--- packages/web/app/components/climb-info/climb-info.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-type ClimbInfo = Record<string, never>;
-
-const ClimbInfo: React.FC<ClimbInfo> = () => {
-  return (
-    <>
-      <h1>TODO: Implement</h1>
-    </>
-  );
-};
-
-export default ClimbInfo;
